전송방식 수정
a태그 POST방식으로 전송되도록 수정

+++ Bowwow/src/main/java/ga/bowwow/service/user/VO/UserAccount.java
@@ -10,7 +10,7 @@ import lombok.ToString;
 @NoArgsConstructor @AllArgsConstructor
 @ToString
 public class UserAccount {
-	long member_serial;
+	Integer member_serial;
 	String id;
 	String email;
 	String citizen_id;
